[WIP] [skip ci] Add General, Compute Capacity and Cli resuable components
In order to get the VCH Configure workflow ready, and since configure workflow does almost the same as the creation workflow (except that reconfigure can patch single groups of information) this commit will introduce several changes to the General and Compute Capacity components in order to make them reusables on Configure View (read only) / VCH Configure Modal (edit mode) and VCH Creation Wizzard.
This will also add an independent CLI Component to be used on any step of the workflow.
This new components (General, Compute Capacity and Cli) are used both on VCH Configure and VCH Creation workflows and also on the Configure view.